Fast API/requirement.txt
fastapi
uvicorn[standard]
tensorflow  
pillow
numpy